+++ src/Hyyan/WPI/Plugin.php
@@ -11,6 +11,8 @@
 namespace Hyyan\WPI;
 
 use Hyyan\WPI\Tools\FlashMessages;
+use Hyyan\WPI\Admin\Settings;
+use Hyyan\WPI\Admin\Features;
 
 /**
  * Plugin.
@@ -191,6 +193,10 @@ class Plugin
         new Shipping();
         new Breadcrumb();
         new Tax();
+        
+        if ('on' === Settings::getOption('localenumbers', Features::getID(), 'on')) {
+            new LocaleNumbers();
+        }        
     }
 
 	/**